requirements/production.txt
Django >= 1.11, < 2.0; python_version < '3'
Django >= 2.1, < 2.2; python_version >= '3'
kobo>=0.4.3
# requires: openldap-devel
python-ldap
djangorestframework >= 3.7.0, < 3.8.0
Markdown
django-filter >= 1.1.0, < 2.0; python_version < '3'
django-filter >= 2.0; python_version >= '3'
django-mptt >= 0.8.6
django-cors-headers >= 2.0.0
djangorestframework-composed-permissions
productmd>=1.2
